+++ lib/overcommit/plugins/commit_msg/single_line_subject.rb
@@ -1,0 +1,13 @@
+module Overcommit::GitHook
+  class SingleLineSubject < HookSpecificCheck
+    include HookRegistry
+
+    def run_check
+      unless user_commit_message[1].to_s.strip.empty?
+        return :warn, 'Subject should be a single line'
+      end
+
+      :good
+    end
+  end
+end